requirements.txt
openai
goose3
chromadb
langchain
PyPDF2
python-dotenv
tiktoken
streamlit